Feature/jasmine unit

+++ src/wwwroot/app/services/vacancyService.js
@@ -17,8 +17,8 @@ export default class VacancyService {
    }
 
    saveVacancy(entity) {
-      if (entity.Id !== undefined) {
-         var additionalUrl = VACANCY_URL + entity.Id;
+      if (entity.id !== undefined) {
+         var additionalUrl = VACANCY_URL + entity.id;
          return _HttpService.put(additionalUrl, entity);
       } else {
          return _HttpService.post(VACANCY_URL, entity);
@@ -26,7 +26,11 @@ export default class VacancyService {
    }
 
    deleteVacancy(entity) {
-      var additionalUrl = VACANCY_URL + entity.Id;
-      _HttpService.remove(additionalUrl, entity);
+      if (entity.id === undefined) {
+         throw new Error('Id should be specified');
+      } else {
+         var additionalUrl = VACANCY_URL + entity.id;
+         _HttpService.remove(additionalUrl, entity);
+      }
    }
 }
